requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytesseract
opencv-python-headless
requests
numpy
pydantic
python-dotenv
google-generativeai
